+++ class/command/AbilityCommand.hpp
@@ -9,7 +9,7 @@ class AbilityCommand : public Command {
     public:
         AbilityCommand(string, string);                     // user defined ctor
         
-        virtual void execute(TableCard&, List<Player>& listPlayer, int playeridx, int& prize, CardsBank& cardsbank, AbilityCardsBank& abilitycardsbank, bool& ascending);
+        virtual void execute(TableCard&, List<Player>& listPlayer, int playeridx, long long int& prize, CardsBank& cardsbank, AbilityCardsBank& abilitycardsbank, bool& ascending);
         virtual bool continueToNextPlayer(List<Player>& listPlayer, int playeridx, AbilityCardsBank& abilitycardsbank);
 
         void doesPlayerHaveAbilityCard(List<Player>& listPlayer, int playeridx, string cardName);
