+++ class/command/Half.hpp
@@ -9,7 +9,7 @@ class Half : public OrdinaryCommand {
     public:
         Half(string, string);           // user defined ctor
 
-        void execute(TableCard&, List<Player>&, int, int&, CardsBank&, AbilityCardsBank&, bool&);
+        void execute(TableCard&, List<Player>&, int, long long int&, CardsBank&, AbilityCardsBank&, bool&);
         bool continueToNextPlayer(List<Player>& listPlayer, int playeridx, AbilityCardsBank& abilitycardsbank);
 };
