+++ class/command/MyCard.cpp
@@ -7,7 +7,7 @@ bool MyCard::continueToNextPlayer(List<Player>& listPlayer, int playeridx, Abili
     return false;
 }
 
-void MyCard::execute(TableCard& tablecard, List<Player>& listPlayer, int playeridx, int& prize, CardsBank& cardsbank, AbilityCardsBank& abilitycardsbank, bool& ascending){
+void MyCard::execute(TableCard& tablecard, List<Player>& listPlayer, int playeridx, long long int& prize, CardsBank& cardsbank, AbilityCardsBank& abilitycardsbank, bool& ascending){
     cout << "---------------------------------------" <<endl;
     cout << "        == " << "Player " << playeridx << "'s Cards ==        " <<endl;
     listPlayer[playeridx-1].seeCard();
