+++ class/command/Switch.cpp
@@ -6,7 +6,7 @@ Switch::Switch(string type, string name) : AbilityCommand(type, name){
 
 }   
 
-void Switch::execute(TableCard& tablecard, List<Player>& listPlayer, int playeridx, int& prize, CardsBank& cardsbank, AbilityCardsBank& abilitycardsbank, bool& ascending) {
+void Switch::execute(TableCard& tablecard, List<Player>& listPlayer, int playeridx, long long int& prize, CardsBank& cardsbank, AbilityCardsBank& abilitycardsbank, bool& ascending) {
     int choice;
     try{
         doesPlayerHaveAbilityCard(listPlayer, playeridx, "Switch");
